Refactor few methods and minor updates in unit tests.
